+++ external-crates/move/crates/move-package-alt/src/cli/mod.rs
@@ -4,10 +4,8 @@
 
 mod build;
 mod new;
-mod parse;
 mod update_deps;
 
 pub use build::Build;
 pub use new::New;
-pub use parse::Parse;
 pub use update_deps::UpdateDeps;
