+++ external-crates/move/crates/move-package-alt/src/git/cache.rs
@@ -58,12 +58,6 @@ pub struct GitTree {
     path_to_repo: PathBuf,
 }
 
-impl Default for GitCache {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
 impl GitCache {
     pub fn new() -> Self {
         Self {
